+++ func/__init__.py
@@ -1,0 +1,3 @@
+# -*- encoding: utf-8 -*-
+# __init__.py
+"""The functions"""